dyn_stack

Struct PodStack

pub struct PodStack { /* private fields */ }

Stack wrapper around a buffer of bytes.

Implementations

impl PodStack

pub fn new(buffer: &mut [u8]) -> &mut Self

Returns a new PodStack from the provided memory buffer.

pub fn new_any<T: Pod>(buffer: &mut [T]) -> &mut Self

Returns a new MemStack from the provided memory buffer.

pub fn can_hold(&self, alloc_req: StackReq) -> bool

Returns true if the stack can hold an allocation with the given size and alignment requirements.

pub fn len_bytes(&self) -> usize

Returns the number of bytes that this stack can hold.

pub fn as_ptr(&self) -> *const u8

Returns a pointer to the stack memory.

pub fn make_aligned_raw<T: Pod>( &mut self, size: usize, align: usize, ) -> (&mut [T], &mut Self)

Returns a new aligned and uninitialized slice and a stack over the remainder of the buffer.

Panics

Panics if the stack isn’t large enough to allocate the array.

pub unsafe fn make_aligned_unpod( &mut self, size: usize, align: usize, ) -> (UnpodStack<'_>, &mut Self)

Returns a new aligned and uninitialized slice and a stack over the remainder of the buffer.

Panics

Panics if the stack isn’t large enough to allocate the array.

Safety

The return value must be dropped if any uninitialized values are written to the bytes by the time the borrow ends.

pub fn make_aligned_with<T: Pod>( &mut self, size: usize, align: usize, f: impl FnMut(usize) -> T, ) -> (&mut [T], &mut Self)

Returns a new aligned slice, initialized with the provided function, and a stack over the remainder of the buffer.

Panics

Panics if the stack isn’t large enough to allocate the array, or if the provided function panics.

pub fn make_raw<T: Pod>(&mut self, size: usize) -> (&mut [T], &mut Self)

Returns a new uninitialized slice and a stack over the remainder of the buffer.

Panics

Panics if the stack isn’t large enough to allocate the array.

pub fn make_with<T: Pod>( &mut self, size: usize, f: impl FnMut(usize) -> T, ) -> (&mut [T], &mut Self)

Returns a new slice, initialized with the provided function, and a stack over the remainder of the buffer.

Panics

Panics if the stack isn’t large enough to allocate the array, or if the provided function panics.

pub fn collect_aligned<I: Pod>( &mut self, align: usize, iter: impl IntoIterator<Item = I>, ) -> (&mut [I], &mut Self)

Returns a new aligned slice, initialized with the provided iterator, and a stack over the remainder of the buffer.
If there isn’t enough space for all the iterator items, then the returned array only contains the first elements that fit into the stack.

Panics

Panics if the provided iterator panics.

pub fn collect<I: Pod>( &mut self, iter: impl IntoIterator<Item = I>, ) -> (&mut [I], &mut Self)

Returns a new slice, initialized with the provided iterator, and a stack over the remainder of the buffer.
If there isn’t enough space for all the iterator items, then the returned array only contains the first elements that fit into the stack.

Panics

Panics if the provided iterator panics.