dyn_stack

Struct StackReq

pub struct StackReq { /* private fields */ }

Stack allocation requirements.

Implementations

impl StackReq

pub const EMPTY: Self

Allocation requirements for an empty unaligned buffer.

pub const OVERFLOW: Self

Unsatisfiable allocation requirements.

pub const fn empty() -> StackReq

Allocation requirements for an empty unaligned buffer.

pub const fn new_aligned<T>(n: usize, align: usize) -> StackReq

Allocation requirements sufficient for n elements of type T, overaligned with alignment align.

Errors

if align is smaller than the minimum required alignment for an object of type T.

if align is not a power of two.

if the size computation overflows

pub const fn new<T>(n: usize) -> StackReq

Allocation requirements sufficient for n elements of type T.

Errors

if the size computation overflows

pub const fn size_bytes(&self) -> usize

The number of allocated bytes required, aligned to self.align_bytes().

pub const fn align_bytes(&self) -> usize

The alignment of allocated bytes required, or 0 if the size overflowed.

pub const fn unaligned_bytes_required(&self) -> usize

The number of allocated bytes required, with no alignment constraints, or usize::MAX in the case of overflow.

Panics

if the size computation overflowed

pub const fn layout(self) -> Result<Layout, LayoutError>

Returns the corresponding layout for the allocation size and alignment.

pub const fn and(self, other: StackReq) -> StackReq

The required allocation to allocate storage sufficient for both of self and other, simultaneously and in any order.

Panics

if the allocation requirement computation overflows.

pub const fn all_of(reqs: &[Self]) -> Self

The required allocation to allocate storage sufficient for all the requirements produced by the given iterator, simultaneously and in any order.

Panics

if the allocation requirement computation overflows.

pub const fn or(self, other: StackReq) -> StackReq

The required allocation to allocate storage sufficient for either of self and other, with only one being active at a time.

Panics

if the allocation requirement computation overflows.

pub fn any_of(reqs: &[StackReq]) -> StackReq

The required allocation to allocate storage sufficient for any of the requirements produced by the given iterator, with at most one being active at a time.

Panics

if the allocation requirement computation overflows.

pub const fn array(self, n: usize) -> StackReq

Same as StackReq::and repeated n times.